+++ packages/atom.io/src/internal/index.ts
@@ -5,6 +5,7 @@ import type { FamilyMetadata } from ".."
 export * from "./get"
 export * from "./set"
 export * from "./is-default"
+export * from "./atom-internal"
 export * from "./selector-internal"
 export * from "./store"
 export * from "./subscribe-internal"
