+++ packages/atom.io/src/transaction.ts
@@ -1,12 +1,5 @@
 import type { ReadonlyValueToken, StateToken } from "."
-import { getState, setState } from "."
-import type { Store } from "./internal"
-import { IMPLICIT } from "./internal"
-import {
-  abortTransaction,
-  finishTransaction,
-  startTransaction,
-} from "./internal/transaction-internal"
+import { transaction__INTERNAL } from "./internal"
 
 export type ƒn = (...parameters: any[]) => any
 
@@ -26,28 +19,8 @@ export type TransactionOptions<ƒ extends ƒn> = {
   do: Action<ƒ>
 }
 
-export const transaction = <ƒ extends ƒn>(
-  options: TransactionOptions<ƒ>,
-  store: Store = IMPLICIT.STORE
-): ((...parameters: Parameters<ƒ>) => ReturnType<ƒ>) & { key: string } =>
-  Object.assign(
-    (...parameters: Parameters<ƒ>) => {
-      startTransaction(store)
-      try {
-        const result = options.do(
-          {
-            get: (token) => getState(token, store),
-            set: (token, value) => setState(token, value, store),
-          },
-          ...parameters
-        )
-        finishTransaction(store)
-        return result
-      } catch (thrown) {
-        abortTransaction(store)
-        store.config.logger?.error(`Transaction ${options.key} failed`, thrown)
-        throw thrown
-      }
-    },
-    { key: options.key }
-  )
+export function transaction<ƒ extends ƒn>(
+  options: TransactionOptions<ƒ>
+): ((...parameters: Parameters<ƒ>) => ReturnType<ƒ>) & { key: string } {
+  return transaction__INTERNAL(options)
+}
